#13 refactor: user 서비스를 command, query, auth로 분리, 인터페이스 추가
