+++ src/main/java/UMC_7th/Closit/domain/user/controller/UserAuthController.java
@@ -5,7 +5,7 @@ import UMC_7th.Closit.domain.user.dto.LoginRequestDto;
 import UMC_7th.Closit.domain.user.dto.RegisterResponseDto;
 import UMC_7th.Closit.domain.user.dto.UserDto;
 import UMC_7th.Closit.domain.user.service.UserAuthService;
-import UMC_7th.Closit.domain.user.service.UserService;
+import UMC_7th.Closit.domain.user.service.UserCommandService;
 import UMC_7th.Closit.global.apiPayload.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -21,12 +21,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserAuthController {
 
-    private final UserService userService;
+    private final UserCommandService userCommandService;
     private final UserAuthService userAuthService;
 
     @PostMapping("/register")
     public ApiResponse<RegisterResponseDto> register (@RequestBody @Valid UserDto userDto){
-        RegisterResponseDto responseDto = userService.registerUser(userDto);
+        RegisterResponseDto responseDto = userCommandService.registerUser(userDto);
 
         return ApiResponse.onSuccess(responseDto);
     }
