+++ src/main/java/UMC_7th/Closit/domain/user/service/UserAuthServiceImpl.java
@@ -1,0 +1,37 @@
+package UMC_7th.Closit.domain.user.service;
+
+import UMC_7th.Closit.domain.user.dto.JwtResponse;
+import UMC_7th.Closit.domain.user.dto.LoginRequestDto;
+import UMC_7th.Closit.domain.user.entity.User;
+import UMC_7th.Closit.domain.user.repository.UserRepository;
+import UMC_7th.Closit.security.jwt.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserAuthServiceImpl implements UserAuthService {
+
+    private final UserRepository userRepository;
+    private final JwtTokenProvider jwtTokenProvider;
+    private final PasswordEncoder passwordEncoder;
+
+    // login
+    @Override
+    public JwtResponse login(LoginRequestDto loginRequestDto) {
+        User user = userRepository.findByEmail(loginRequestDto.getEmail())
+                .orElseThrow(() -> new IllegalArgumentException("가입되지 않은 이메일입니다."));
+
+        if (!passwordEncoder.matches(loginRequestDto.getPassword(), user.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+
+        String accessToken = jwtTokenProvider.createAccessToken(user.getEmail(), null);
+        String refreshToken = jwtTokenProvider.createRefreshToken(user.getEmail(), null);
+
+        return new JwtResponse(accessToken, refreshToken);
+    }
+
+}
