+++ src/main/java/UMC_7th/Closit/domain/user/service/UserQueryService.java
@@ -1,0 +1,4 @@
+package UMC_7th.Closit.domain.user.service;
+
+public interface UserQueryService {
+}
